Extract ModuleMapType into separate file
